NO AUTO Strips graph, lock, cypher, path, search procs from full
