+++ core/src/main/java/apoc/graph/Graphs.java
@@ -1,6 +1,6 @@
 package apoc.graph;
 
-import apoc.cypher.Cypher;
+import apoc.cypher.CypherUtils;
 import apoc.graph.document.builder.DocumentToGraph;
 import apoc.graph.util.GraphsConfig;
 import apoc.result.RowResult;
@@ -16,6 +16,7 @@ import org.neo4j.procedure.*;
 
 import java.util.*;
 import java.util.stream.Stream;
+import static apoc.graph.GraphsUtils.extract;
 
 /**
  * @author mh
@@ -38,37 +39,6 @@ public class Graphs {
         Set<Relationship> rels = new HashSet<>(10000);
         extract(data, nodes,rels);
         return Stream.of(new VirtualGraph(name,nodes,rels,properties));
-    }
-
-    public static boolean extract(Object data, Set<Node> nodes, Set<Relationship> rels) {
-        boolean found = false;
-        if (data == null) return false;
-        if (data instanceof Node) {
-            nodes.add((Node)data);
-            return true;
-        }
-        else if (data instanceof Relationship) {
-            rels.add((Relationship) data);
-            return true;
-        }
-        else if (data instanceof Path) {
-            Iterables.addAll(nodes,((Path)data).nodes());
-            Iterables.addAll(rels,((Path)data).relationships());
-            return true;
-        }
-        else if (data instanceof Iterable) {
-            for (Object o : (Iterable)data) found |= extract(o,nodes,rels);
-        }
-        else if (data instanceof Map) {
-            for (Object o : ((Map)data).values()) found |= extract(o,nodes,rels);
-        }
-        else if (data instanceof Iterator) {
-            Iterator it = (Iterator) data;
-            while (it.hasNext()) found |= extract(it.next(), nodes,rels);
-        } else if (data instanceof Object[]) {
-            for (Object o : (Object[])data) found |= extract(o,nodes,rels);
-        }
-        return found;
     }
 
     @Description("apoc.graph.fromPath(path,'name',{properties}) - creates a virtual graph object for later processing")
@@ -102,7 +72,7 @@ public class Graphs {
         Set<Node> nodes = new HashSet<>(1000);
         Set<Relationship> rels = new HashSet<>(1000);
         Map<String,Object> props = new HashMap<>(properties);
-        tx.execute(Cypher.withParamMapping(statement, params.keySet()), params).stream().forEach(row -> {
+        tx.execute(CypherUtils.withParamMapping(statement, params.keySet()), params).stream().forEach(row -> {
             row.forEach((k,v) -> {
                 if (!extract(v,nodes,rels)) {
                     props.put(k,v);
