+++ core/src/main/java/apoc/path/LabelMatcher.java
