+++ core/src/main/java/apoc/path/LabelMatcherGroup.java
