+++ core/src/main/java/apoc/path/LabelSequenceEvaluator.java
