+++ core/src/main/java/apoc/path/NodeEvaluators.java
