+++ core/src/main/java/apoc/path/Paths.java
