+++ core/src/main/java/apoc/path/RelationshipSequenceExpander.java
