+++ core/src/main/java/apoc/search/ParallelNodeSearch.java
