+++ full/src/test/java/apoc/ApocSplitTest.java
@@ -17,38 +17,7 @@ import static org.junit.Assert.assertTrue;
  */
 public class ApocSplitTest {
     public static final List<String> PROCEDURES_FROM_CORE = List.of(
-        "apoc.util.sleep",
-        "apoc.util.validate",
         "apoc.example.movies",
-        "apoc.graph.fromData",
-        "apoc.graph.from",
-        "apoc.graph.fromPath",
-        "apoc.graph.fromPaths",
-        "apoc.graph.fromDB",
-        "apoc.graph.fromCypher",
-        "apoc.graph.fromDocument",
-        "apoc.graph.validateDocument",
-        "apoc.lock.all",
-        "apoc.lock.nodes",
-        "apoc.lock.read.nodes",
-        "apoc.lock.rels",
-        "apoc.lock.read.rels",
-        "apoc.cypher.runTimeboxed",
-        "apoc.cypher.run",
-        "apoc.cypher.runMany",
-        "apoc.cypher.runManyReadOnly",
-        "apoc.cypher.doIt",
-        "apoc.cypher.runWrite",
-        "apoc.cypher.runSchema",
-        "apoc.when",
-        "apoc.do.when",
-        "apoc.case",
-        "apoc.do.case",
-        "apoc.search.nodeAllReduced",
-        "apoc.search.nodeReduced",
-        "apoc.search.multiSearchReduced",
-        "apoc.search.nodeAll",
-        "apoc.search.node",
         "apoc.load.jsonArray",
         "apoc.load.json",
         "apoc.load.jsonParams",
@@ -62,22 +31,8 @@ public class ApocSplitTest {
         "apoc.help");
 
     public static final List<String> FUNCTIONS_FROM_CORE = List.of(
-        "apoc.util.sha1",
-        "apoc.util.sha256",
-        "apoc.util.sha384",
-        "apoc.util.sha512",
-        "apoc.util.md5",
-        "apoc.util.validatePredicate",
-        "apoc.util.decompress",
-        "apoc.util.compress",
-        "apoc.path.create",
-        "apoc.path.slice",
-        "apoc.path.combine",
-        "apoc.path.elements",
         "apoc.label.exists",
         "apoc.diff.nodes",
-        "apoc.cypher.runFirstColumnMany",
-        "apoc.cypher.runFirstColumnSingle",
         "apoc.xml.parse",
         "apoc.version",
         "apoc.scoring.existence",
